listener.py payload and broker changes...
The MQTT payload is now in JSON to make parsing it more readable.
Listener can grab the topic from the MQTT message.
Listener subscribes to the topic: 'plux/data/+'
Listener tries to connect to the unencrypted but authenticated MQTT broker that was setup by @figomez.

NOTE: The username and password on line 25 are not the actual username and password.
      Those are just placeholders. I didn't want to push code with a password in it.
      Check Slack #general for username & password.